Split up course guide specs
